+++ docs/CHANGELOG.md
@@ -95,10 +95,12 @@ extra bounds from `AnimatablePrefab` and `AnimationSetPrefab` ([#1435])
 [#1504]: https://github.com/amethyst/amethyst/pull/1504
 
 ### Removed
+
 - Removed all `NetEvent's` because they were not used. ([#1539])
 - Removed filter logic, because it didn't do anything, will be added back in a later version (NetFilter, FilterConnected). ([#1539])
 
 ### Fixed
+
 * Optimize loading of wavefront obj mesh assets by getting rid of unnecessary allocations. ([#1454])
 * Fixed the "json" feature for amethyst_assets. ([#1302])
 * Fixed default system font loading to accept uppercase extension ("TTF"). ([#1328])
@@ -106,6 +108,7 @@ extra bounds from `AnimatablePrefab` and `AnimationSetPrefab` ([#1435])
 * Fix omission in `PosNormTangTex` documentation. ([#1371])
 * Fix division by zero in vertex data building ([#1481])
 * Fix tuple index generation on `PrefabData` and `EventReader` proc macros. ([#1501])
+* Avoid segmentation fault on Windows when using `AudioBundle` in `amethyst_test`. ([#1595], [#1599])
 
 [#1114]: https://github.com/amethyst/amethyst/pull/1114
 [#1213]: https://github.com/amethyst/amethyst/pull/1213
@@ -165,6 +168,8 @@ extra bounds from `AnimatablePrefab` and `AnimationSetPrefab` ([#1435])
 [#1584]: https://github.com/amethyst/amethyst/pull/1584
 [#1591]: https://github.com/amethyst/amethyst/pull/1591
 [#1582]: https://github.com/amethyst/amethyst/pull/1582
+[#1595]: https://github.com/amethyst/amethyst/issues/1595
+[#1599]: https://github.com/amethyst/amethyst/pull/1599
 
 ## [0.10.0] - 2018-12